+++ packages/emoji/src/constants.ts
@@ -1,0 +1,5 @@
+// v13 isnt available on platforms yet
+export const MAX_EMOJI_VERSION = 12.1;
+
+// v5 is for emoji version v13 (above)
+export const LATEST_DATASET_VERSION = '4.2.1';